Add BrowseObjectBase implementation
This type was previously provided by the Project System as a workaround to the
fact that the shell's didn't have a method as virtual. That was changed in
https://dev.azure.com/devdiv/DevDiv/_git/VS/pullrequest/248337, and the type
made internal in https://github.com/dotnet/project-system/pull/6198.
Once the updated shell is available, `BrowseObjectBase` can be removed and
`LocalizableProperties` used directly in its place.