Catch ProjectException when querying properties (#3494)
Under a rare set of circumstances where a project multi-targets and a given reference is not defined in the active configuration (described in dotnet/project-system#6311) it's possible that an item is not known by the project.

This change catches the exception to prevent larger failures in the tree. Without this, the node appears expanded (its triangle points down) but no children are displayed beneath.

(cherry picked from commit 1cea7b7479bbb59081078481edc567f2c219fd5c)