Fix error when a project doesn't not specify a restore output path
If a project does not import Microsoft.Common.props, it won't have the properties set to determine the restore output path.  This can cause us to crash when we call `Path.Combine()` with a null value.

This fix checks for this condition, logs a message, and leaves the project out of restore.