Add metadata to filter MEF exports until needed (#3671)
Previously the import of IAttachedCollectionSourceProvider parts did not filter exports based on metadata. This meant that all exports were loaded, which resulted in DLLs such as NuGet.Client being loaded even when the dependencies tree was unused.

This filtering is now supported by Solution Explorer, and this commit adds the relevant condition to ensure DLLs are not loaded unnecessarily.