Make package icon loading in PM UI async to prepare for codespaces improvements needed. (#3702)
When connected to a codespaces sesssion, we are going to need to call the server instance to get embedded license files and embedded icons (for packages not accessed via search to http feeds).

In order to make that work well, we need to change our Icon fetching to be more async.

This change doesn't add the codespaces side of things yet...is just trying to make our current icon loading as good, and better than it is now.

The ViewModel.IconUrl property now has a matching IconBitmap property...and IconBitmapStatus

Since 2016, we've used a ValueConverter (runs on UI thread only) to convert the Url to a BitmapImage that an Image element could show. This approach causes a UI delay when network retrieval of images is slow. We believe this work should fix that.

In order to move off the UI thread...we moved the download into background work that IconBitmap getter kicks off. We rely on INotifyPropertyChanged to notify databinding later, that the IconBitmap value has changed on the background thread. Binding can update UI properties on UI Thread that were modified on a viewmodel in a background thread(s).

ErrorFloodGate.cs - made this thread safe.

Testing/Validation
Tests Added: Yes
Validation: testing with instrumented version of UI to see where each image came from (memorycache, download, default, defaultduetonetworkproblem, etc...). tested with clumsy to slow network, drop packets, etc...