INuGetPackageFileService for icons and licenses (#3823)
Fixes: https://github.com/NuGet/Home/issues/10151 and https://github.com/NuGet/Home/issues/10080
Introduces a service to get icons and licenses. When running in standalone mode, all works in same process. When running in codespaces mode, can be remoted across process and machine boundaries.
All fetches of icons and licenses (from file:, https:, or embedded in an already downloaded package), are performed on the server...and the client gets the resultant stream.
[PR of spec](https://github.com/NuGet/Home/pull/10460)