Remove assignments to MSBuildAllProjects
In the past, MSBuildAllProjects was used to support incremental build. Every project file would have to append its path to this property, and an up-to-date check could use that information to scan timestamps of all consituent files.

Nowadays, MSBuild tracks which of the project files has the newest timestamp and automatically prepends that value to MSBuildAllProjects. The up-to-date check in VS only needs to check the first item from this property, and the rest are ignored.

For more information see http://www.panopticoncentral.net/2019/07/12/msbuildallprojects-considered-harmful/