Handle deep link URI and opens PMUI from an open solution  (#4207)
Handles NuGet deep link URI's and parses said URI for validity. From an open solution in Visual Studio, PMUI loads in and searches for a target NuGet package based off the information provided in the URI. If a NuGet package version is specified, it will look for a package with said version.