Fix package content search in Solution Explorer
Sometime ago the target framework identifier used in the lock file was changed from form `.NETCoreApp,Version=v5.0` to `net5.0` (presumably to allow for platform suffixes such as `net5.0-android`). This broke a lookup within the dependencies tree in Solution Explorer, leading to results not being displayed.

This change allows the value to match the project's `TargetFramework` property as well, which fixes this issue on the projects I tested against.