Fix IsSubSetOfWithNullCheck (#11222)
IsSubSetOfWithNullCheck is supposed to return true when `other` is
a subset of `parent`.

This fixes an issue where restore can take a long time to run on
solutions with many references and packages.

A test solution was created with 42 projects, with project n referencing
projects 1..(n-1). Each project also had a couple of package references
and a NoWarn element.

Before this change, restore ran for 5 minutes, used ~14GB of memory,
and failed with an OutOfMemoryException ("Array dimensions exceeded
supported range.")

After this change, restore completed in under 5 seconds.