Merged PR 427043: Fix build failures in 4.9.x branch
Fix build failures. New Migration framework code access `Environment.GetFolderPath` method and `Environment.SpecialFolder.LocalApplicationData` API which are not available in netstandard1.6. Skipped migration logic for netstandard1.6 because of this reason. It does not impact because this is a nuget.exe release that can only run on net472 on Windows and Mono.