Display documents from packages in the dependencies tree
Packages may contain files such as README files, license text, etc. This change adds a "Documents" node beneath packages in the dependencies tree of Solution Explorer, when such files exist, allowing them to be opened and viewed in the editor.

It's also possible to search in Solution Explorer for these items, so long as the "Search within external items" option is checked.