Avoid trying hitting thread pool immediately after solution load
NuGet initialization hits VsMachineWideSettings.Settings immediately after solution load on the UI thread and blocks on it. This leads to be sensitive to high thread pool latency. Instead, if the UI thread is already blocked on us, just read the settings on the UI thread - its doing absolutely nothing else anyway.,