Use default '/' split symbol in ETW events.
Avoid converting event/property names in ETW provider

This is allocating a lot during solution close and other places where lots of events are being produced, which is causing GCs to occur in some runs, introducing noise. The statement that says ETW events can't have '/' seems to be incorrect, as Windows, VS & .NET all include events with '/'. Also stopped converting properties as these are raw strings.

This reduces ~300-400ms of noise in a time of 1500 second scenario in Solution Close for speedometer.