Reduce size of LibraryDependency class
The `LibraryDependency` class has been identified as contributing to high GC on customer machines. This change reduces the memory footprint of that type in two ways:

1. Six integral fields are packed into a single 32-bit integer, reducing 12 bytes down to 4 bytes.
2. The backing collection for the `NoWarn` property is now lazily allocated, as it will usually be empty.

For the packing of fields, there are three Boolean and three enum-based fields. With some masking and shifting, these can all be efficiently stored within a single 32-bit integer space. Unit tests have been added to ensure the correctness of those fields now and in future. Comments have been added to the enums being packed to ensure any future modifications to those enums take this packing strategy into account. The packing of enums does leave some headroom, so a few more members could be added to each without requiring any code changes in `LibraryDependency`.

For the `NoWarn` collection, an additional `NoWarnCount` property was added that allows callers to test whether the collection is empty before calling the `NoWarn` getter (which would lazily allocate a backing collection). This seems to be a pragmatic approach to this issue given the backwards compatibility requirements. Several callers have been updated to check the count first.