Reduce allocations in TargetFrameworkInformation.Clone (#5199)
The previous clone operation had two allocation-related issues:

1. The property initializers would allocate collections during object contruction, only for them to be replaced during population of the clone.
2. The cloning of properties didn't take advantage of known collection sizes when constructing new collections, leading to both:
  - Temporary containers and extra work during collection resizing.
  - Oversized collections upon construction, where extra unused capacity remains in the collections, wasting space.

This change creates a "clone constructor" which takes responsibility for constructing the clone such that each object is only constructed once, and in an efficient manner. This addresses the above problems.

This method was identified as a top contributor to GC pauses by GCPauseWatson.