Create SingleFileProvider
Previously, `ContentFileUtils.GetContentFileGroup` would construct a `VirtualFileProvider` class, passing only a single string to the constructor. That class performs a bunch of string chopping and multiple collection allocations. It then loops through them, performing a bunch of string operations and string concatenations in order to produce results within the `GetDirectoryContents` method.

This change adds a `SingleFileProvider` implementation of `IFileProvider` that's specialised for this single-file case. It doesn't need to do any string splits, and only retains the original string. When `GetDirectoryContents` is called, it uses `ReadOnlySpan<char>` to avoid allocating temporary strings. It also keeps a singleton instance of the "none found" result, to avoid allocating anything in such cases.

Unit tests have been added that cover both implementations when a single string exists, to ensure behavioral parity between implementations.