Reduce allocations in RuntimeDescription and RuntimeDependencySet
These two classes follow a similar construction and clone pattern, and the changes to each are similar.

Previously each expected to receive non-null enumerables with input values, which would then be enumerated to materialise copies.

In cases where these collections were empty, the allocations are avoidable.

This changes the constructors to receive nullable values, allowing singleton collections to be used in such cases.

Similarly, when cloning, if the collections are empty, nulls are passed so that these singletons are applied. Also when cloning, the known collection type is cast to to enable allocation-free enumeration of the source when constructing the clone object.