Avoid dictionary lookups in loops
Several places were looping through a dictionary's keys, then looking up each value by that key within the loop. This is inneficient as both the key and value are availble for enumeration, by enumerating the dictionary directly.

This pattern is particularly troublesome for large collections, and for `ConcurrentDictionary<>` which can incur a lot of overhead from locking when contention exists.

Another gain here is that the `Keys` property lazily allocates a `Dictionary<>.KeyCollection` instance, so these changes avoid that allocation too.

We add a `Deconstruct` extension method to make enumeration of dictionaries a little more ergonomic as well, allowing deconstructing directly into two local variables within the `foreach` statement itself. I had previously added `DeconstructionExtensions` for use in dependencies tree code, and this change moves it to the shared folder for use elsewhere in the solution.