Configure custom spelling dictionary (#5262)
This adds a `spelling.dic` file to the solution and configured it as the place to store custom words that should be ignored by the spell checker. This helps to keep the number of IDE diagnostics down, making it easier to focus on real issues.

Also addresses a few typos in some files I had been working on.