Reduce allocations in NuGet.DependencyResolver.Tracker
This object keeps track of graph items as they are observed. The internal implementation has been optimised to pack data in a more compact fashion, as this type was identified as contributing to GC pauses in VS by GCPauseWatson.

Unit tests were added to document the previous behaviour, then the class was rewritten to use less memory.

Some of the optimisations:
- Don't create an entry for every look up operation. Only create it when some state actually needs to be written (adding a graph item, or marking ambiguous).
- Don't allocate the HashSet for every entry in the tracker. Only allocate one once there is more than one item for a given library.
- Don't call Contains on the HashSet before calling Add. If the item already exists, Add is a no-op.