Don't allocate temporaries in FrameworkNameProvider.GetVersionString (#5250)
Previously this would allocate a stack, enumerators, and a copy of the stack (for a call to Linq's `Reverse`).

With this change all state required to produce the version string is kept on the stack.

The unit test coverage for this method has been increased.

Also the allocation of an empty `Version` was removed in the `Try*` methods when they return false, to make them match their signatures and avoid a wasted allocation. Consuming code in this repo has been reviewed, and the behaviour captured in unit tests.