Reduce allocations in VersionRangeFormatter (#5279)
`VersionRangeFormatter` calls into `VersionFormatter` for almost all of its operations. The latter produces a `string` which the former then composes into its output. This intermediary `string` is a wasteful allocation that can be avoided. GCPauseWatson shows that `VersionRangeFormatter` calls into `VersionFormatter` often enough that these temporary strings contribute to GC pauses on customer machines in the real world.

This change creates a pathway for `VersionRangeFormatter` to pass a `StringBuilder` to `VersionFormatter` in order to avoid the temporary string allocation. It also changes `VersionRangeFormatter` to use a `StringBuilder` more consistently internally, and to pass that to other methods it invokes where possible.