Reduce size of LockFileTargetLibrary
This class has many fields of `List<T>`, most of which will remain empty throughout the lifetime of the class. Having a field for each of these takes space, and so does the empty list allocated for each.

This commit removes the fields and replaces them with a sparse dictionary such that only those properties that are actually populated will be allocated in memory.

We also introduce the concept of "freezing" this object, after which accessing properties will no longer return a lazily constructed mutable `List<T>`, and instead will return `Array.Empty<T>` which cannot be modified. This prevents code that reads these properties from inadvertently materializing these lazy collections.

In order to use a `readonly struct record` on .NET Framework, a reference to the source package `IsExternalInit` is added.

Null annotations are added to `LockFileTargetLibrary`.