Correctly identify transitive package references
Fixes https://github.com/NuGet/Home/issues/10368

When a transitive package dependency has the same identity as a direct
project dependency that is differenced by a different inner build (TFM),
the dependency is incorrectly resolved as a project.

The resolver logic invoked by the dependency walker only knows about which
projects are referenced but not by which TFM. To correctly isolate inner
builds from each other and avoid incorrectly promoting a project
dependency, pass the nearest project dependencies to the dependency
resolving logic.