Fix missing transitive dependency tree nodes in some multi-targeting projects
Fixes https://github.com/dotnet/project-system/issues/6832

The "target name" used in the project model can take different forms. For example, both `net8.0` and `.NETFramework,Version=v4.8.1` are valid values. The dependencies tree uses the "target alias" for its target nodes, in multi-target projects. This is because two different aliases can resolve to the same TFM, for example, yet attached items may differ.

This change digs the alias out of the lock file and uses that value throughout. It also renames a property and fixes the documentation on it so as to avoid confusion.