Fix hang in Solution Explorer search through transitive package items of dependencies tree (#5917)
* Revert "Reduce number of Dataflow updates through dependency tree code (#5508)"

This reverts commit d85c6ea4588c5b4f2b4a4d7bf8073837fb308f58.

Turns out that the search provider uses `GetLatestVersionAsync` on this
data source, meaning it must propagate a value for every input version.
I will fix the original issue in a different way, by updating the
consumer to ignore version-only changes instead.

* Avoid tree updates when the value is unchanged

This replaces the approach taken in #5508 in a way that doesn't interfere with Solution Explorer search.