Display warnings/errors on transitive package/project references
Historically we have not shown warning/error icons on transitive package/project reference nodes in the tree. Since the introduction of NuGet Audit, it becomes more important to add these icons, and propagate them up the spine of the tree in order to guide the user from the "Dependencies" node down to the specific node, however deep within the tree.

The implementation builds a map from node to ancestors then walks from nodes with warnings up the tree, remembering the highest severity seen at each level. We then use that value on each node to select an appropriate icon for the log level.

Note that this change does not set icons on top-level dependencies. These come from design-time build data and will need to be handled separately by changes to NuGet's MSBuild tasks/targets.