Make PackageSpecDependencyProvider consider the version being requested
Fixes: https://github.com/NuGet/Home/issues/10368
Re-submit of https://github.com/NuGet/NuGet.Client/pull/5452

NuGet gets confused with transitive project and package
dependencies with the same identity (name) in a multi-targeting project
and selects the wrong type (project instead of package).

Projects are expected to be preferred over packages, but only when the
version matches. If the version doesn't match, projects shouldn't be
selected in frameworks which don't declare a ProjectReference to those
projects.