Add version check to DotnetSign test for int assembly
Modified DotnetSign_SignPackageWithInsecureCertificateFingerprint_RaisesWarningAsync in DotnetSignTests.cs to include a conditional check based on the major version of the int assembly. If the major version is 10 or higher, the test asserts that the signing result is unsuccessful and that the errors contain the insecure certificate fingerprint code. If the major version is less than 10, the test retains the original behavior, asserting that the signing result is successful and that the output contains the insecure certificate fingerprint code.