Reduce allocations under UnresolvedMessages.GetMessageAsync
This method shows as 0.6% (95 MB) of all VS allocations from VS start until opening the file in the CSharpEditingTest speedometer test. I think the changes in this PR should get rid of the majority of those allocations.

The primary change here is to not use a SortedSet, but instead have the code do the sorting when it's needed. The SortedSet ctor allocates a fair bit, and each node in the set requires an allocation for the wrapping node.

As for the concern of now needing to walk the whole collection in GetBestMatch as the collection is no longer sorted, the old code was doing that anyways. Both the FirstOrDefault and Last methods are IEnumerable extensions that potentially walk the full collection.