Support building on Linux when full signing is not available
Full signing requires the platform to provide the RSA+SHA1 algorithms.
This is not true for some Linux distributions such as Fedora 42, or RHEL
9 (and 10). This was done to match the RSA+SHA1 deprecation:

- https://devblogs.microsoft.com/devops/ssh-rsa-deprecation/
- https://www.openssh.com/txt/release-8.2

It's easy to work around this by setting PublicSign=true and
DelaySign=false, however that's not easy to do without code
modifications. With this change, a user can set properties at the build
command line. The following lets me build and test nuget-client on
Fedora 42 without issues.

    $ ./build.sh -p:PublicSign=true -p:DelaySign=false